+++ test.py
@@ -6,7 +6,7 @@ import yaml
 import argparse
 from detectron2.layers.rotated_boxes import pairwise_iou_rotated
 
-from lib.post_process import post_process
+from lib.general import post_process
 from lib.load import load_data
 from lib.loss import ComputeLoss
 from lib.logger import logger
